Adds SignupList component
