+++ client/src/Components/UserDetail/UserDetail.js
@@ -3,94 +3,12 @@ import { Col, Panel, Image, Label, ListGroup, ListGroupItem, PageHeader, Row, Ta
 import Moment from 'react-moment';
 import PropTypes from 'prop-types';
 import MessageList from '../MessageList/MessageListContainer';
+import SignupList from '../SignupList/SignupListContainer';
 import VotingPlan from './VotingPlan';
 
 const queryString = require('query-string');
 const helpers = require('../../helpers');
 const config = require('../../config');
-
-function postLabel(status) {
-  let style = 'warning';
-  if (status === 'rejected') {
-    style = 'danger';
-  } else if (status === 'accepted') {
-    style = 'success';
-  }
-  return <Label bsStyle={style}>{status}</Label>;
-}
-
-function renderSignup(signup) {
-  let posts = null;
-  const numPosts = signup.posts.data.length;
-  if (numPosts) {
-    posts = signup.posts.data.map((post, index) => {
-      const postDate = <Moment format={'MM/DD/YY'}>{post.created_at}</Moment>;
-      let postSource = null;
-      if (post.source) {
-        postSource = ` via ${helpers.formatSource(post.source)}`;
-      }
-      const status = postLabel(post.status);
-
-      let whyParticipated = null;
-      // Posts are ordered by ascending created date. We only ask for Why Participated if it's
-      // the User's first post for the Campaign.
-      if (index === 0) {
-        whyParticipated = (
-          <ListGroupItem>
-            <strong>Why Participated:</strong> {signup.why_participated}
-          </ListGroupItem>
-        );
-      }
-      return (
-        <ListGroup key={post.id}>
-          <ListGroupItem>
-            <Image src={post.media.url} height={200} />
-          </ListGroupItem>
-          <ListGroupItem>
-            <strong>Caption:</strong> {post.media.caption}
-          </ListGroupItem>
-          <ListGroupItem>
-            <strong>Quantity:</strong> {signup.quantity}
-          </ListGroupItem>
-          {whyParticipated}
-          <ListGroupItem>
-            <strong>Submitted:</strong> {postDate}{postSource} {status}
-          </ListGroupItem>
-        </ListGroup>
-      );
-    });
-  }
-  const campaignId = signup.campaign_id;
-  const campaignLink = <a href={`/campaigns/${campaignId}`}>{campaignId}</a>;
-  const source = signup.signup_source ? ` via ${helpers.formatSource(signup.signup_source)}` : null;
-  return (
-    <tr key={signup.signup_id}>
-      <td><strong>{campaignLink}</strong></td>
-      <td>
-        <a href={signup.url}><Moment format={'MM/DD/YY'}>{signup.created_at}</Moment>{source}</a>
-      </td>
-      <td>{posts}</td>
-    </tr>
-  );
-}
-
-function renderSignups(signups) {
-  const rows = signups.map(signup => renderSignup(signup));
-  return (
-    <Table>
-      <thead>
-        <tr>
-          <th width={120}>Campaign</th>
-          <th width={200}>Joined</th>
-          <th>Posts</th>
-        </tr>
-      </thead>
-      <tbody>
-        {rows}
-      </tbody>
-    </Table>
-  );
-}
 
 function userInfo(user) {
   const lastMessagedDate = <Moment format={config.dateFormat}>{ user.last_messaged_at }</Moment>;
@@ -179,7 +97,7 @@ function tabs(user) {
   const numConversations = Object.keys(user.conversations).length;
   const signupsTab = (
     <Tab eventKey={numConversations + 1} title={signupsLabel}><br />
-      { renderSignups(user.signups.data) }
+      <SignupList signups={user.signups.data} />
     </Tab>
   );
   const activeKey = platform ? 1 : 0;
